Last Update on Apr 27, 2026: Welcome to the Agentic Era with MagicAI: Introducing GPT 5.5, GPT Image 2.0, Opus 4.7, Gemini 3.1 Flash Live, Lyria 3 Pro, Skills (Paid), Deep Research (Paid), AI Council Mode (Paid), GPT 5.4-Mini, GPT 5.4-Nano, Redesigned AI Chat Pro Experience, Smart Highlight, Chat Navigator, Redesigned Sidebar, Smart Image Display, All-new External Chatbot v3 and Inbox, Sales Agent for External Chatbot (Paid), Booking Agent for External Chatbot (Paid), Voice Chat for External Chatbot (Paid), Feedback for External Chatbot, Customer Segmentation for External Chatbot, Instagram for External Chatbot (Paid), GPT-5.3 Instant, GPT-5.4, Social Media Agent (Paid), AI Blogger Agent (Paid), Follow up Question for AI Chat Pro, Ability to Publish Stories with Social Media Suite & Agent, AI Integration for Creative Suite, Carousel Posts & Smart Prompts for Social Media, Youtube & Youtube Shorts Integration for Social Media, Nano Banana 2, Opus 4.6, Sonnet 4.6, Gemini 3.1 Pro, Kling 3, Kling 2.6, Kling Motion Control, Grok Imagine, ElevenLabs v3, AI Image Pro (Paid), Folders for AI Chat Pro, Memory for AI Chat Pro, AI Fashion Studio (Paid), Performance Boost, GPT 5.2, GPT 5.2 Pro, Gemini 3 Flash, GPT-Image 1.5, Gemini 3 Flash, Gemini 3 Pro, GPT 5.1, NanoBanana Pro, Grok 4.1 and more, AI Presentation Maker (Paid), Veo 3.1, Claude Sonnet 4.5, Claude Opus 4.1, Kling AI 2.5 Turbo, AI Music Pro (Paid), Cover and Emoji Support for AI Editor, Chat with Documents for AI Chat Pro, GPT-5 Pro, Sora 2, Sora 2 Pro, Seedream v4, Brand-New External Chatbot and Inbox, Facebook Messenger for External Chatbot (Paid), Multi-Model AI (Paid) and more..
Last Update on Apr 27, 2026: Welcome to the Agentic Era with MagicAI: Introducing GPT 5.5, GPT Image 2.0, Opus 4.7, Gemini 3.1 Flash Live, Lyria 3 Pro, Skills (Paid), Deep Research (Paid), AI Council Mode (Paid), GPT 5.4-Mini, GPT 5.4-Nano, Redesigned AI Chat Pro Experience, Smart Highlight, Chat Navigator, Redesigned Sidebar, Smart Image Display, All-new External Chatbot v3 and Inbox, Sales Agent for External Chatbot (Paid), Booking Agent for External Chatbot (Paid), Voice Chat for External Chatbot (Paid), Feedback for External Chatbot, Customer Segmentation for External Chatbot, Instagram for External Chatbot (Paid), GPT-5.3 Instant, GPT-5.4, Social Media Agent (Paid), AI Blogger Agent (Paid), Follow up Question for AI Chat Pro, Ability to Publish Stories with Social Media Suite & Agent, AI Integration for Creative Suite, Carousel Posts & Smart Prompts for Social Media, Youtube & Youtube Shorts Integration for Social Media, Nano Banana 2, Opus 4.6, Sonnet 4.6, Gemini 3.1 Pro, Kling 3, Kling 2.6, Kling Motion Control, Grok Imagine, ElevenLabs v3, AI Image Pro (Paid), Folders for AI Chat Pro, Memory for AI Chat Pro, AI Fashion Studio (Paid), Performance Boost, GPT 5.2, GPT 5.2 Pro, Gemini 3 Flash, GPT-Image 1.5, Gemini 3 Flash, Gemini 3 Pro, GPT 5.1, NanoBanana Pro, Grok 4.1 and more, AI Presentation Maker (Paid), Veo 3.1, Claude Sonnet 4.5, Claude Opus 4.1, Kling AI 2.5 Turbo, AI Music Pro (Paid), Cover and Emoji Support for AI Editor, Chat with Documents for AI Chat Pro, GPT-5 Pro, Sora 2, Sora 2 Pro, Seedream v4, Brand-New External Chatbot and Inbox, Facebook Messenger for External Chatbot (Paid), Multi-Model AI (Paid) and more..

+++ resources/views/default/panel/admin/dashboard/finance.blade.php
@@ -92,7 +92,7 @@
                     <span class="font-medium text-heading-foreground">{{ __('Total Spending') }}</span>
                     <h3 class="lqd-statistic-change m-0 flex items-center text-2xl sm:text-[30px]">
                         <span
-                            class="text-xl">{{ $currencySymbol }}</span>{{ number_format(setting('total_spend', 0), 1) }}
+                            class="text-xl">{{ $currencySymbol }}</span>{{ number_format( (int) setting('total_spend', 0), 1) }}
                     </h3>
                 </div>
             </x-card>
